Relocate Programs component
